backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pytest==7.4.0
cryptography==41.0.7 